refactor: moved strategies files to subfolder
